feat(matrix): add 2 handling of matrix for managed and unmanaged study

+++ antarest/study/storage/rawstudy/model/filesystem/lazy_node.py
@@ -18,7 +18,7 @@ from zipfile import ZipFile
 
 from typing_extensions import override
 
-from antarest.matrixstore.matrix_uri_mapper import MatrixUriMapper
+from antarest.matrixstore.matrix_uri_mapper import MatrixUriMapper, MatrixUriMapperManaged
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.inode import G, INode, S, V
 
@@ -129,9 +129,13 @@ class LazyNode(INode, ABC, Generic[G, S, V]):  # type: ignore
         self._assert_url_end(url)
 
         if isinstance(data, str) and self.matrix_mapper.matrix_exists(data):
-            self.get_link_path().write_text(data)
-            if self.config.path.exists():
-                self.config.path.unlink()
+            if isinstance(self.matrix_mapper, MatrixUriMapperManaged):
+                self.get_link_path().write_text(data)
+                if self.config.path.exists():
+                    self.config.path.unlink()
+            else:
+                matrix = self.matrix_mapper.get_matrix(data)
+                self.dump(cast(S, matrix))
             return None
 
         self.dump(cast(S, data), url)
